refactor: use field decorator
