+++ packages/varlet-ui/src/option/option.less
@@ -1,7 +1,7 @@
 :root {
   --option-height: 38px;
   --option-padding: 0 12px;
-  --option-selected-background: var(--input-decorator-focus-color);
+  --option-selected-background: var(--field-decorator-focus-color);
 }
 
 .var-option {
